Fix homebrew instructions
The instructions for installing a specific version of solidity using homebrew no longer work with the latest version of homebrew. There is a specific error message that indicates that raw github links are forbidden. This documents the workaround that is now needed. This is how I did it but of course I am open so suggestions.