Update cheatsheet.rst
The documentation currently says:

````
``abi.encodeCall(function functionPointer, (...)) returns (bytes memory)``: ABI-encodes a call to ``functionPointer`` with the arguments found in the
  tuple. Performs a full type-check, ensuring the types match the function signature. Result equals ``abi.encodeWithSelector(functionPointer.selector, (...))``
````

However, this is not correct.

`abi.encodeWithSelector` does not take a tuple as a second argument, instead, it takes an unended list of arguments. Saying these two are identical, is then not correct:

- `abi.encodeCall(function functionPointer, (...)) `
- `abi.encodeWithSelector(functionPointer.selector, (...))`

It should be that these two are identical:

- `abi.encodeCall(function functionPointer, (...)) `
- `abi.encodeWithSelector(functionPointer.selector, ...)`

subtle, but maybe there is a clearly way to say that `encodeWithSelector` is the same as the unpacked tuple from `encodeCall`.